Add code comments and make code path clearer. Remove packing check as packing support for pretokenised data is merged to trl. See https://github.com/huggingface/trl/pull/2011
